+++ src/screens/account_details/components/native_account/types.ts
@@ -2,18 +2,6 @@ export type OverviewType = {
   address: string;
 }
 
-export type AccountType = {
-  address: string;
-}
-
-export type AccountsType = {
-  data: {
-    [value:number]: AccountType[];
-  };
-  count: number;
-  loading: boolean;
-}
-
 export type BalanceType = {
   native: TokenUnit;
   stake: TokenUnit;