+++ CHANGELOG.md
@@ -26,4 +26,5 @@
 - Add volume on token page ([\#112](https://github.com/forbole/big-dipper-2.0-solana/issues/112))
 - Fix tabs index behavior (validators page) ([\#69](https://github.com/forbole/big-dipper-2.0-solana/issues/69))
 - Add native account details page
+- Add nonce account details page